requirements.txt
Django==4.2.7
Pillow==10.1.0
python-decouple==3.8
pyOpenSSL==24.0.0
cryptography==41.0.7
openpyxl==3.1.2
pandas==2.1.4
requests>=2.32.0
zeep>=4.3.0
lxml>=4.6.0
xmlsec>=1.3.13
cryptography>=41.0.0
